Add my name (Mengué) as stopword for Pod::Spell
The auto process that adds author names as stopwords is not enough.
It looks like the copyright/licence text added by Pod::Weaver takes
the copyright owner from somewhere else than the dist.ini, and so
in the POD my name has an accent. But that accent is not in the
author list in dist.ini used to add stopwords in t/author-pod-spell.t.
So I'm adding my name manually in "=for stopwords" in POD.